requirements.txt
black==22.6.0
django==4.2
djangorestframework==3.14.0
django-cors-headers==3.14.0
